refactorizando y evitando duplicacion en manejo de errores
